% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{load_mat}
\alias{load_mat}
\title{Load a matrix from a file}
\usage{
load_mat(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Load a matrix from a file
}
